login before conversation requests

+++ server/src/main/java/com/friendsync/muush/controller/ConversationController.java
@@ -14,9 +14,10 @@ import com.friendsync.muush.repo.TDO.CreateConversationRequest;
 import com.friendsync.muush.service.ConversationService;
 
 import jakarta.annotation.Resource;
-import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
-import lombok.Getter;
+import jakarta.servlet.http.HttpSession;
+
+import static com.friendsync.stevenpang.constant.UserConstant.USER_LOGIN_STATE;
 
 @RestController
 @RequestMapping("/conversation")
@@ -28,6 +29,10 @@ public class ConversationController {
     public ResponseEntity<?> searchConversation(
         HttpServletRequest request,
         @RequestParam(value = "keyword", required = false, defaultValue = "") String keyword) {
+            HttpSession session = request.getSession();
+            if (session == null || session.getAttribute(USER_LOGIN_STATE) == null)
+                return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("login first");
+
             if (!keyword.equals("")) {
                 return ResponseEntity.status(HttpStatus.OK).body(service.searchConversations(keyword));
             } else {
@@ -40,6 +45,9 @@ public class ConversationController {
         HttpServletRequest request,
         @RequestBody CreateConversationRequest createConversationRequest
     ) {
+        HttpSession session = request.getSession();
+        if (session == null || session.getAttribute(USER_LOGIN_STATE) == null)
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("login first");
         Conversation c = service.createConversation(
             createConversationRequest.getOwnerId(),
             createConversationRequest.getInformation(),
@@ -52,10 +60,13 @@ public class ConversationController {
 
     @GetMapping("/generate")
     public ResponseEntity<?> genLicense(
-        HttpServlet request,
+        HttpServletRequest request,
         @RequestParam(value = "id") Long id,
         @RequestParam(value = "owner") Long owner
     ) {
+        HttpSession session = request.getSession();
+        if (session == null || session.getAttribute(USER_LOGIN_STATE) == null)
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("login first");
         String s = service.generateLicense(id, owner);
         if (s != null)
             return ResponseEntity.ok(s);
@@ -64,12 +75,27 @@ public class ConversationController {
     }
 
     @GetMapping("/getown")
-    public ResponseEntity<?> getOwn(@RequestParam("owner") Long owner) {
+    public ResponseEntity<?> getOwn(
+        HttpServletRequest request,
+        @RequestParam("owner") Long owner
+    ) {
+        HttpSession session = request.getSession();
+        if (session == null || session.getAttribute(USER_LOGIN_STATE) == null)
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("login first");
+        
+        
         return ResponseEntity.ok(service.getOwnConversations(owner));
     }
 
     @GetMapping("/delete")
-    public ResponseEntity<?> delete(@RequestParam("id") Long id, @RequestParam("owner") Long owner) {
+    public ResponseEntity<?> delete(   
+        HttpServletRequest request,
+        @RequestParam("id") Long id,
+        @RequestParam("owner") Long owner
+     ) {
+        HttpSession session = request.getSession();
+        if (session == null || session.getAttribute(USER_LOGIN_STATE) == null)
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("login first");
         boolean result = service.deleteConversation(id, owner);
         if (result)
             return ResponseEntity.ok("success");
